Encode the progress bar state in the type system
Applying with_delims to a unbounded iterator does not create a delimiter
around the characters. At the same time, this will not display any
error to the user. "The core issue is that progress is now a stateful
object, based on whether or not it is in bounded or unbounded state. And
the methods are allowed to be called regardless of in which state you
are in" (from the original author in https://youtu.be/bnnacleqg6k).

Here we are applying Type State to help solve the issue.